QD_functions split into qd_ligand_opt & qd_dissociate
