+++ src/qmflows/components/__init__.py
@@ -5,9 +5,11 @@ from .reactivity import (Distance, Angle, Dihedral)
 from .qd_database import (read_database, compare_database, write_database)
 from .qd_import_export import (read_mol, set_prop, create_dir)
 from .qd_ams import (check_sys_var, qd_opt, ams_job_mopac_sp)
-from .qd_functions import (optimize_ligand, find_substructure, find_substructure_split,
+from .qd_functions import (find_substructure, find_substructure_split, get_time,
                            rotate_ligand, merge_mol, qd_int, adf_connectivity, fix_h,
                            fix_carboxyl, update_coords)
+from .qd_ligand_opt import optimize_ligand
+from .qd_dissociate import get_topology_dict, dissociate_ligand
 
 __all__ = [
     'ADF3FDE_Result', 'Fragment', 'MFCC_Result', 'adf3fde', 'adf_fragmentsjob', 'mfcc',
@@ -15,5 +17,7 @@ __all__ = [
     'read_database', 'compare_database', 'write_database',
     'read_mol', 'set_prop', 'create_dir',
     'check_sys_var', 'qd_opt', 'ams_job_mopac_sp',
-    'optimize_ligand', 'find_substructure', 'find_substructure_split', 'rotate_ligand',
-    'merge_mol', 'qd_int', 'adf_connectivity', 'fix_h', 'fix_carboxyl', 'update_coords']
+    'find_substructure', 'find_substructure_split', 'rotate_ligand',
+    'merge_mol', 'qd_int', 'adf_connectivity', 'fix_h', 'fix_carboxyl', 'update_coords',
+    'optimize_ligand',
+    'get_topology_dict', 'dissociate_ligand']
