+++ src/qmflows/components/qd_ams.py
@@ -2,7 +2,6 @@ __all__ = ['check_sys_var', 'ams_job_mopac_sp', 'qd_opt']
 
 import os
 import shutil
-import time
 
 from scm.plams import (Settings, AMSJob, init, finish, Units)
 from scm.plams.tools.kftools import KFFile
@@ -11,10 +10,7 @@ from scm.plams.interfaces.adfsuite.scmjob import (SCMJob, SCMResults)
 import scm.plams.interfaces.molecule.rdkit as molkit
 
 from .qd_import_export import export_mol
-from .qd_functions import (adf_connectivity, fix_h, fix_carboxyl)
-
-
-time_print = '[' + time.strftime('%H:%M:%S') + '] '
+from .qd_functions import (adf_connectivity, fix_h, fix_carboxyl, get_time)
 
 
 def check_sys_var():
@@ -25,13 +21,13 @@ def check_sys_var():
     sys_var_exists = [item in os.environ for item in sys_var]
     for i, item in enumerate(sys_var_exists):
         if not item:
-            print(time_print +
+            print(get_time() +
                   'WARNING: The environment variable ' + sys_var[i] + ' has not been set')
     if False in sys_var_exists:
-        raise EnvironmentError(time_print + 'One or more ADF environment variables have '
+        raise EnvironmentError(get_time() + 'One or more ADF environment variables have '
                                'not been set, aborting ADF job.')
     if '2018' not in os.environ['ADFHOME']:
-        error = time_print + 'No ADF version 2018 detected in ' + os.environ['ADFHOME']
+        error = get_time() + 'No ADF version 2018 detected in ' + os.environ['ADFHOME']
         error += ', aborting ADF job.'
         raise ImportError(error)
